+++ instrumentation/opentelemetry-instrumentation-aiohttp-client/src/opentelemetry/instrumentation/aiohttp_client/__init__.py
@@ -262,7 +262,9 @@ def _instrument(
     url_filter: _UrlFilterT = None,
     request_hook: _RequestHookT = None,
     response_hook: _ResponseHookT = None,
-    trace_configs: typing.Optional[aiohttp.TraceConfig] = None,
+    trace_configs: typing.Optional[
+        typing.Sequence[aiohttp.TraceConfig]
+    ] = None,
 ):
     """Enables tracing of all ClientSessions
 
@@ -270,16 +272,15 @@ def _instrument(
     the session's trace_configs.
     """
 
-    if trace_configs is None:
-        trace_configs = []
+    trace_configs = trace_configs or ()
 
     # pylint:disable=unused-argument
     def instrumented_init(wrapped, instance, args, kwargs):
         if context_api.get_value(_SUPPRESS_INSTRUMENTATION_KEY):
             return wrapped(*args, **kwargs)
 
-        if kwargs.get("trace_configs"):
-            trace_configs.extend(kwargs.get("trace_configs"))
+        client_trace_configs = list(kwargs.get("trace_configs", ()))
+        client_trace_configs.extend(trace_configs)
 
         trace_config = create_trace_config(
             url_filter=url_filter,
@@ -288,9 +289,9 @@ def _instrument(
             tracer_provider=tracer_provider,
         )
         trace_config._is_instrumented_by_opentelemetry = True
-        trace_configs.append(trace_config)
+        client_trace_configs.append(trace_config)
 
-        kwargs["trace_configs"] = trace_configs
+        kwargs["trace_configs"] = client_trace_configs
         return wrapped(*args, **kwargs)
 
     wrapt.wrap_function_wrapper(